Update CQC providers API call to use new endpoint and move into CQC API file
