+++ backend/server/utils/CQCDataAPI.js
@@ -12,4 +12,17 @@ module.exports = {
 
     return data;
   },
+
+  getCQCProviderData: async function (locationID) {
+    const regex = new RegExp('^[0-9 -]{1,}$', 'gm');
+    let isCorrectFormat = regex.test(locationID);
+
+    if (isCorrectFormat) {
+      const { data } = await axios.get(cqcEndpoint + '/providers/' + locationID, {
+        headers: { 'Ocp-Apim-Subscription-Key': cqcSubscriptionKey },
+      });
+      return data;
+    }
+    return {};
+  },
 };
